+++ build.rs
@@ -1,6 +1,5 @@
 fn main() {
     winrt_macros::build!(
         foundation
-        types
     );
 }
